pyproject.toml
[project]
name = "airplay-streamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyatv>=0.16.1",
    "pyyaml>=6.0.3",
]
